requirements.txt
pandas==0.25.3
yfinance==0.1.54
plotly==4.9.0
GetOldTweets3==0.0.11
Flask~=1.1.2
utils~=1.0.1
PyYAML~=5.3.1
tweepy~=3.8.0
numpy~=1.19.0
cassandra-driver~=3.24.0
pytz~=2020.1
iso8601~=0.1.13
gunicorn
Flask-Cors=3.0.9